Accept reality that System.in is not interruptable
In our unique case that is. This means ConsoleReaderThread will -always- be alive for the length of the application but no commands will be parsed without Input running. This is an acceptable solution that fixes the problem stemming from the 9 year old bug found here:
http://bugs.sun.com/bugdatabase/view_bug.do?bug_id=4514257

Thanks to @kitskub with slamming me with reality.

Signed-off-by: Chris Sanders <zidane@outlook.com>